requirements.txt
streamlit>=1.30.0
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.4.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
pandas>=2.0.0
openpyxl>=3.1.0
Pillow>=10.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
fpdf2>=2.7.0
firebase-admin>=6.0.0
google-cloud-firestore>=2.0.0
